fix: drop the composite-expression cast in the relay coil clear
Remove the (uint8_t) cast from the coil mask built in
set_relay_coils_low(). The masked-write conversion in 712589c introduced
it, and it fails the MISRA gate: `make STRICT_TOOLS=1 analyze-misra`
reports misra-c2012-10.8 at src/bypass_output_tq2_l2_5v_relay.c:51,
"the value of a composite expression shall not be cast to a different
essential type category or a wider essential type".

The cast is not required and the finding is not a real violation. The
mask is a constant expression whose essential type is already unsigned
char, so casting it to uint8_t neither changes category nor widens.
cppcheck 2.13 reaches the opposite conclusion because it resolves the
cast's own value type to unsigned int -- 16 bits under
--platform=avr8 -- while typing the composite correctly as unsigned
char, and the addon then compares 16 against 8 and reports a widening
cast. Instrumenting misra.py against the dump shows both values
directly. The mis-resolution depends on surrounding translation-unit
context and does not reproduce in an isolated file with the same
include set, so there is nothing to pin an inline suppression to.

Deleting the cast is preferred over waiving the rule. The D-1 deviation
scopes 10.8 to the two MCU shells where avr-libc's _SFR_* macros force
composite casts that cannot be written any other way; the modular
output drivers are outside that boundary by design, and widening the
waiver to cover a cast the code does not need would weaken it. The
uncast form also matches how output_mask is already built a few lines
above in the same function, and it costs nothing: every shipping image
is byte-identical to the pre-change build.

Hoisting the mask into a uint8_t const local is equally MISRA-clean but
was rejected -- XC8 places the const local in flash, costing one word
on PIC10F322, which has 19 of 512 words free.

Local verification:
- make STRICT_TOOLS=1 analyze-misra: clean
- flash unchanged on every shell carrying the driver:
  ATtiny13a 864 B, ATtiny202 994 B, PIC10F320 245 w,
  PIC10F322 493 w, PIC12F675 563 w
- builds clean under -Werror -Wall -Wextra -Wconversion (avr-gcc)
  and XC8 v3.10
- make STRICT_TOOLS=1 test-long on a clean copy of this tree:
  132 mutants killed / 0 survived, mutation sandbox 131 checks,
  golden-model coverage 99.39%, verified-core coverage 100.00%,
  release preflight/provenance/qualification 52/83/48 checks

Only the classic analyze-misra lane scans the output drivers; the
pic10f322, attiny202 and pic10f320 MISRA lanes each scan their own
shell translation unit, so no other gate was affected.

Provisioned qualification required after push:
- make STRICT_TOOLS=1 test-long
- scripts/ci-local.sh

+++ src/bypass_output_tq2_l2_5v_relay.c
@@ -48,8 +48,8 @@ void hw_init_output_pins(void) {
 
 // force both coils low
 static void set_relay_coils_low(void) {
-    hw_pin_mask_set_low((uint8_t)(
-            (1U << RELAY_RESET_PIN) | (1U << RELAY_SET_PIN)));
+    hw_pin_mask_set_low(
+            (1U << RELAY_RESET_PIN) | (1U << RELAY_SET_PIN));
 }
 
 void hw_set_bypass_state(void) {
